LayoutTests/fast/webgpu/regression/repro_290633-expected.txt

CONSOLE MESSAGE: no validation error
CONSOLE MESSAGE: Pass
